refactor: move stream handler into sequencing/consensus
