+++ crates/sequencing/papyrus_consensus/src/lib.rs
@@ -11,6 +11,8 @@ pub mod simulation_network_receiver;
 pub mod single_height_consensus;
 #[allow(missing_docs)]
 pub mod state_machine;
+#[allow(missing_docs)]
+pub mod stream_handler;
 #[cfg(test)]
 pub(crate) mod test_utils;
 #[allow(missing_docs)]
